PS_2_Cognitive_Radiology_Report/requirements.txt
transformers
timm
pandas
numpy
opencv-python
albumentations
Pillow
pydicom
nltk
rouge-score
scikit-learn
matplotlib
seaborn
tqdm
kaggle
radgraph
kagglehub
setuptools
